refactor: reorganize oauth-user-consent-flow package structure and add integration test configuration
