+++ core/python/oauth-user-consent-flow/app/agent_engine_app.py
@@ -21,9 +21,9 @@ from google.adk.artifacts import GcsArtifactService, InMemoryArtifactService
 from google.cloud import logging as google_cloud_logging
 from vertexai.agent_engines.templates.adk import AdkApp
 
-from adk_ae_oauth.agent import app as adk_app
-from adk_ae_oauth.app_utils.telemetry import setup_telemetry
-from adk_ae_oauth.app_utils.typing import Feedback
+from app.agent import app as adk_app
+from app.app_utils.telemetry import setup_telemetry
+from app.app_utils.typing import Feedback
 
 # Load environment variables from .env file at runtime
 load_dotenv()
