+++ core/python/oauth-user-consent-flow/app/tools.py
@@ -26,7 +26,7 @@ from google.auth.transport.requests import Request
 from google.oauth2.credentials import Credentials
 from googleapiclient.discovery import build
 
-from adk_ae_oauth import auths
+from app import auths
 
 logger = logging.getLogger(__name__)
